+++ lib/CircularModulesPlugin.js
@@ -5,18 +5,23 @@
 
 "use strict";
 
-/** @typedef {import("../Module")} Module */
-/** @typedef {import("../ModuleGraph")} ModuleGraph */
+/** @typedef {import("./Compiler")} Compiler */
+/** @typedef {import("./ModuleGraph")} ModuleGraph */
+/** @typedef {import("./Module")} Module */
+/** @typedef {import("./Module").BuildInfo} BuildInfo */
+
+const PLUGIN_NAME = "CircularModulesPlugin";
 
 /**
  * Detects circular dependencies among synchronous module imports.
  *
- * Builds an adjacency list from module.dependencies (excluding async blocks),
- * then runs an iterative SCC algorithm to find circular modules.
+ * Builds a compressed-sparse-row adjacency layout from module.dependencies
+ * (excluding async blocks), then runs an iterative SCC algorithm to find
+ * circular modules.
  *
  * Use the static `build()` method to create an instance. All intermediate data
- * (adjacency lists, index mappings) is local to `build()` and released on return.
- * The instance only holds the result.
+ * (adjacency layout, index mappings) is local to `build()` and released on
+ * return. The instance only holds the result.
  */
 class CycleGraph {
 	/**
@@ -149,4 +154,29 @@ class CycleGraph {
 	}
 }
 
-module.exports = CycleGraph;
+/**
+ * Detects circular dependencies and marks each circular module
+ * via buildInfo.isCircular for downstream consumers.
+ */
+class CircularModulesPlugin {
+	/**
+	 * @param {Compiler} compiler the compiler instance
+	 * @returns {void}
+	 */
+	apply(compiler) {
+		compiler.hooks.compilation.tap(PLUGIN_NAME, (compilation) => {
+			compilation.hooks.optimizeModules.tap(PLUGIN_NAME, (modules) => {
+				const { circularModules } = CycleGraph.build(
+					modules,
+					compilation.moduleGraph
+				);
+				for (const m of modules) {
+					/** @type {BuildInfo} */
+					(m.buildInfo).isCircular = circularModules.has(m);
+				}
+			});
+		});
+	}
+}
+
+module.exports = CircularModulesPlugin;
