+++ lib/WebpackOptionsApply.js
@@ -63,7 +63,6 @@ const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 const JavascriptParser = require("./javascript/JavascriptParser");
 
 const JsonModulesPlugin = require("./json/JsonModulesPlugin");
-const CircularModulesPlugin = require("./optimize/CircularModulesPlugin");
 
 const ChunkPrefetchPreloadPlugin = require("./prefetch/ChunkPrefetchPreloadPlugin");
 
@@ -392,7 +391,6 @@ class WebpackOptionsApply extends OptionsApply {
 		}
 
 		new JavascriptModulesPlugin().apply(compiler);
-		new CircularModulesPlugin().apply(compiler);
 
 		new JsonModulesPlugin().apply(compiler);
 		new AssetModulesPlugin({
@@ -639,6 +637,12 @@ class WebpackOptionsApply extends OptionsApply {
 			const InnerGraphPlugin = require("./optimize/InnerGraphPlugin");
 
 			new InnerGraphPlugin().apply(compiler);
+		}
+
+		if (options.mode === "production") {
+			const CircularModulesPlugin = require("./CircularModulesPlugin");
+
+			new CircularModulesPlugin().apply(compiler);
 		}
 
 		const ConstExportsPlugin = require("./optimize/ConstExportsPlugin");
